+++ omop_alchemy/toolkit/analytics/adverse_events/__init__.py
@@ -30,11 +30,11 @@ come from — which readings were admitted, over what window — is decided
 in ``omop_alchemy.toolkit.analytics.body_metrics``.
 """
 
-from .ctcae import CTCAEWeightLoss, ctcae_weight_loss_grade
 from .weight_loss import (
     MartinWeightLoss,
     critical_weight_loss_grade,
     martin_weight_loss_grade,
+    CTCAEWeightLoss, ctcae_weight_loss_grade
 )
 
 __all__ = [
